requirements.txt
astroid==2.11.2
certifi==2021.10.8
charset-normalizer==2.0.12
dill==0.3.4
flake8==4.0.1
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.5.1
prompt-toolkit==1.0.14
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
PyInquirer==1.0.3
pylint==2.13.4
python-dotenv==0.20.0
regex==2022.3.15
requests==2.27.1
six==1.16.0
stripe==2.69.0
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
wrapt==1.14.0
